Show one favorite event

+++ app/src/main/java/POJO/User.java
@@ -38,7 +38,7 @@ public class User {
     protected ArrayList<String> friendsList;
 
     @ColumnInfo(name = "favoriteEventIdsList")
-    protected ArrayList<Integer> favoriteEventIdsList;
+    protected ArrayList<Long> favoriteEventIdsList;
 
     @NonNull
     @ColumnInfo(name = "password")
@@ -79,17 +79,17 @@ public class User {
         this.setPreference(preference);
     }
 
-    public ArrayList<Integer> getFavoriteEventIdsList() {
+    public ArrayList<Long> getFavoriteEventIdsList() {
         return favoriteEventIdsList;
     }
 
-    public void setFavoriteEventIdsList(ArrayList<Integer> eventsIdsList) {
+    public void setFavoriteEventIdsList(ArrayList<Long> eventsIdsList) {
         this.favoriteEventIdsList = eventsIdsList;
     }
-    public void addFavoriteEventId(Integer id){
+    public void addFavoriteEventId(Long id){
         this.favoriteEventIdsList.add(id);
     }
-    public void removeFavoriteEventId(Integer id){
+    public void removeFavoriteEventId(Long id){
         this.favoriteEventIdsList.remove(id);
     }
